Rework titlecase algorithm used to give collection pages a default title
The previous algorithm would titlecase “2nd meeting” as “2Nd Meeting” which the new one does not.

OTOH the new one will turn “3d movie” into “3d Movie”.

Since this function is only used to create a nice title from collection pages without one, which mainly is drafts, I opted for simpler code rather than build up a long list of special cases.

If we want better titlecase I think we should externalize the code (rely on a gem) as there will be many exceptions (like abbreviations) so a comprehensive test suite is called for and possibly even a way to customize the list of exceptions, so that it can be extended to multiple languages.